## Description

This is the spike to convert simplefeatures geometries directly into GEOS geometries rather than using WKB.

Unfortunately, the results show that it's not as fast as the existing WKB approach.

Note that the spike is limited only to Polygons.

## Check List

Have you:

- Added unit tests? N/A

- Add cmprefimpl tests? (if appropriate?) N/A

## Related Issue

- #190 

## Benchmark Results

```
name                    old time/op    new time/op    delta
Intersection/n=10-4       48.7µs ± 7%    46.0µs ± 4%   -5.44%  (p=0.005 n=15+11)
Intersection/n=100-4      95.1µs ± 8%   100.8µs ± 9%   +5.95%  (p=0.008 n=15+15)
Intersection/n=1000-4      401µs ±12%     489µs ±10%  +22.02%  (p=0.000 n=15+15)
Intersection/n=10000-4    3.45ms ±10%    4.68ms ±12%  +35.93%  (p=0.000 n=12+14)
NoOp/n=10-4               3.83µs ± 6%    4.05µs ±10%   +5.77%  (p=0.001 n=14+14)
NoOp/n=100-4              12.1µs ± 7%    19.7µs ± 4%  +62.05%  (p=0.000 n=13+14)
NoOp/n=1000-4             90.3µs ± 8%   175.9µs ± 4%  +94.83%  (p=0.000 n=15+13)
NoOp/n=10000-4             926µs ±15%    1712µs ± 6%  +84.91%  (p=0.000 n=15+14)

name                    old alloc/op   new alloc/op   delta
Intersection/n=10-4       1.19kB ± 0%    0.50kB ± 0%  -57.72%  (p=0.000 n=15+15)
Intersection/n=100-4      6.33kB ± 0%    1.48kB ± 0%  -76.61%  (p=0.000 n=14+14)
Intersection/n=1000-4     55.0kB ± 0%    11.2kB ± 0%  -79.61%  (p=0.000 n=15+15)
Intersection/n=10000-4     558kB ± 0%     115kB ± 0%  -79.37%  (p=0.000 n=15+15)
NoOp/n=10-4                 864B ± 0%      408B ± 0%  -52.78%  (p=0.000 n=15+15)
NoOp/n=100-4              5.68kB ± 0%    2.02kB ± 0%  -64.37%  (p=0.000 n=15+15)
NoOp/n=1000-4             49.5kB ± 0%    16.6kB ± 0%  -66.40%  (p=0.000 n=14+15)
NoOp/n=10000-4             492kB ± 0%     164kB ± 0%  -66.64%  (p=0.000 n=14+14)

name                    old allocs/op  new allocs/op  delta
Intersection/n=10-4         31.0 ± 0%      17.0 ± 0%  -45.16%  (p=0.000 n=15+15)
Intersection/n=100-4        31.0 ± 0%      17.0 ± 0%  -45.16%  (p=0.000 n=15+15)
Intersection/n=1000-4       31.0 ± 0%      17.0 ± 0%  -45.16%  (p=0.000 n=15+15)
Intersection/n=10000-4      31.0 ± 0%      17.0 ± 0%  -45.16%  (p=0.000 n=15+15)
NoOp/n=10-4                 21.0 ± 0%      13.0 ± 0%  -38.10%  (p=0.000 n=15+15)
NoOp/n=100-4                21.0 ± 0%      13.0 ± 0%  -38.10%  (p=0.000 n=15+15)
NoOp/n=1000-4               21.0 ± 0%      13.0 ± 0%  -38.10%  (p=0.000 n=15+15)
NoOp/n=10000-4              21.0 ± 0%      13.0 ± 0%  -38.10%  (p=0.000 n=15+15)
```